Consolidate ChildIndex into model for common use
